:sparkles: Feat: API 통신에서 발생하는 응답 및 예외 처리
